// src/context/edit-context/actions.types.ts
export enum EDIT_CONTEXT_ACTIONS {
  EDIT_STATUS = "EDIT_STATUS",
  UPDATE_BRIGHTNESS = "UPDATE_BRIGHTNESS",
  UPDATE_CONTRAST = "UPDATE_CONTRAST",
  UPDATE_SATURATION = "UPDATE_SATURATION",
  UPDATE_GRAYSCALE = "UPDATE_GRAYSCALE",
  UPDATE_ROTATE = "UPDATE_ROTATE",
  UPDATE_ZOOM = "UPDATE_ZOOM",
  UPDATE_HORIZONTAL_FLIP = "UPDATE_HORIZONTAL_FLIP",
  UPDATE_VERTICAL_FLIP = "UPDATE_VERTICAL_FLIP",
  RESET = "RESET",
  SAVE_EDIT = "SAVE_EDIT",
  DOWNLOAD_IMAGE = "DOWNLOAD_IMAGE",
}